api_gateway/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.12.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.104.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
limits==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.2
narwhals==1.44.0
numpy==2.2.6
openai==0.28.0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
plotly==6.2.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydeck==0.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uuid==1.30
uvicorn==0.24.0
wrapt==1.17.2
yarl==1.20.1
